pyproject.toml
[tool.poetry]
name = "nhk_programs-checker"
version = "0.1.0"
description = ""
authors = ["yrarchi <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pyyaml = "^6.0.1"
slack-sdk = "^3.26.0"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
